requirements.txt
tvdb-v4-official==1.1.0
python-dotenv==1.0.1
tvdb-v4-official==1.1.0
colorama==0.4.6